requirements.txt
pydub>=0.25.1
noisereduce>=3.0.0
numpy>=1.24.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
